+++ libs/framework-core/source/ftrack_framework_core/host/__init__.py
@@ -16,7 +16,7 @@ from ftrack_utils.decorators import (
     with_new_session,
     delegate_to_main_thread_wrapper,
 )
-from ftrack_utils.calls.methods import call_directly
+from ftrack_utils.decorators.threading import call_directly
 
 logger = logging.getLogger(__name__)
 
@@ -310,6 +310,7 @@ class Host(object):
         client_options = event['data']['client_options']
         payload = event['data']['payload']
 
+        # TODO: we should be able to replace this to: tool_config = self.registry.get_one(extension_type='tool_config', reference=tool_config_reference)
         for typed_configs in self.tool_configs.values():
             tool_config = None
             for _tool_config in typed_configs: