+++ libs/utils/source/ftrack_utils/decorators/__init__.py
@@ -7,4 +7,5 @@ from ftrack_utils.decorators.track_usage import track_framework_usage
 from ftrack_utils.decorators.threading import (
     run_in_main_thread,
     delegate_to_main_thread_wrapper,
+    call_directly,
 )