+++ libs/utils/release_notes.md
@@ -1,5 +1,8 @@
 # ftrack Utils library release Notes
 
+## upcoming
+
+* [changed] framework tool configs; get_groups method dictionary support on filters argument..
 
 ## v3.0.0
 2024-09-19